Move data_frame_analytics from group1 to own group
