+++ x-pack/test/functional/apps/ml/data_frame_analytics/outlier_detection_creation.ts
@@ -5,20 +5,18 @@
  * 2.0.
  */
 
-import { AnalyticsTableRowDetails } from '../../../../services/ml/data_frame_analytics_table';
-import { FtrProviderContext } from '../../../../ftr_provider_context';
+import { FtrProviderContext } from '../../../ftr_provider_context';
+import { AnalyticsTableRowDetails } from '../../../services/ml/data_frame_analytics_table';
 
 export default function ({ getService }: FtrProviderContext) {
   const esArchiver = getService('esArchiver');
   const ml = getService('ml');
   const editedDescription = 'Edited description';
 
-  describe('classification saved search creation', function () {
+  describe('outlier detection creation', function () {
     before(async () => {
-      await esArchiver.loadIfNeeded('x-pack/test/functional/es_archives/ml/farequote_small');
-      await ml.testResources.createIndexPatternIfNeeded('ft_farequote_small', '@timestamp');
-      await ml.testResources.createSavedSearchFarequoteLuceneIfNeeded('ft_farequote_small');
-      await ml.testResources.createSavedSearchFarequoteKueryIfNeeded('ft_farequote_small');
+      await esArchiver.loadIfNeeded('x-pack/test/functional/es_archives/ml/ihp_outlier');
+      await ml.testResources.createIndexPatternIfNeeded('ft_ihp_outlier', '@timestamp');
       await ml.testResources.setKibanaTimeZoneToUTC();
 
       await ml.securityUI.loginAsMlPowerUser();
@@ -26,43 +24,66 @@ export default function ({ getService }: FtrProviderContext) {
 
     after(async () => {
       await ml.api.cleanMlIndices();
-      await ml.testResources.deleteSavedSearches();
-      await ml.testResources.deleteIndexPatternByTitle('ft_farequote_small');
+      await ml.testResources.deleteIndexPatternByTitle('ft_ihp_outlier');
     });
 
-    const dateNow = Date.now();
+    const jobId = `ihp_1_${Date.now()}`;
+
     const testDataList = [
       {
-        suiteTitle: 'with lucene query',
-        jobType: 'classification',
-        jobId: `fq_saved_search_2_${dateNow}`,
-        jobDescription: 'Classification job based on a saved search with lucene query',
-        source: 'ft_farequote_lucene',
+        suiteTitle: 'iowa house prices',
+        jobType: 'outlier_detection',
+        jobId,
+        jobDescription: 'Outlier detection job based on ft_ihp_outlier dataset with runtime fields',
+        source: 'ft_ihp_outlier',
         get destinationIndex(): string {
-          return `user-${this.jobId}`;
+          return `user-${jobId}`;
         },
         runtimeFields: {
-          uppercase_airline: {
+          lowercase_central_air: {
             type: 'keyword',
-            script: 'emit(params._source.airline.toUpperCase())',
+            script: 'emit(params._source.CentralAir.toLowerCase())',
           },
         },
-        dependentVariable: 'airline',
-        trainingPercent: 20,
-        modelMemory: '20mb',
+        modelMemory: '5mb',
         createIndexPattern: true,
         expected: {
-          source: 'ft_farequote_small',
-          rocCurveColorState: [
-            // tick/grid/axis
-            { color: '#DDDDDD', percentage: 38 },
-            // line
-            { color: '#98A2B3', percentage: 7 },
+          histogramCharts: [
+            { chartAvailable: true, id: '1stFlrSF', legend: '334 - 4692' },
+            { chartAvailable: true, id: 'BsmtFinSF1', legend: '0 - 5644' },
+            { chartAvailable: true, id: 'BsmtQual', legend: '0 - 5' },
+            { chartAvailable: true, id: 'CentralAir', legend: '2 categories' },
+            { chartAvailable: true, id: 'Condition2', legend: '2 categories' },
+            { chartAvailable: true, id: 'Electrical', legend: '2 categories' },
+            { chartAvailable: true, id: 'ExterQual', legend: '1 - 4' },
+            { chartAvailable: true, id: 'Exterior1st', legend: '2 categories' },
+            { chartAvailable: true, id: 'Exterior2nd', legend: '3 categories' },
+            { chartAvailable: true, id: 'Fireplaces', legend: '0 - 3' },
           ],
-          runtimeFieldsEditorContent: ['{', '  "uppercase_airline": {', '    "type": "keyword",'],
+          scatterplotMatrixColorsWizard: [
+            // markers
+            { color: '#52B398', percentage: 15 },
+            // grey boilerplate
+            { color: '#6A717D', percentage: 13 },
+          ],
+          scatterplotMatrixColorStatsResults: [
+            // red markers
+            { color: '#D98071', percentage: 1 },
+            // tick/grid/axis, grey markers
+            { color: '#6A717D', percentage: 12 },
+            { color: '#D3DAE6', percentage: 8 },
+            { color: '#98A1B3', percentage: 12 },
+            // anti-aliasing
+            { color: '#F5F7FA', percentage: 30 },
+          ],
+          runtimeFieldsEditorContent: [
+            '{',
+            '  "lowercase_central_air": {',
+            '    "type": "keyword",',
+          ],
           row: {
             memoryStatus: 'ok',
-            type: 'classification',
+            type: 'outlier_detection',
             status: 'stopped',
             progress: '100',
           },
@@ -71,65 +92,15 @@ export default function ({ getService }: FtrProviderContext) {
               {
                 section: 'state',
                 expectedEntries: {
-                  id: `fq_saved_search_2_${dateNow}`,
+                  id: jobId,
                   state: 'stopped',
                   data_counts:
-                    '{"training_docs_count":320,"test_docs_count":1284,"skipped_docs_count":0}',
-                  description: 'Classification job based on a saved search with lucene query',
+                    '{"training_docs_count":1460,"test_docs_count":0,"skipped_docs_count":0}',
+                  description:
+                    'Outlier detection job based on ft_ihp_outlier dataset with runtime fields',
                 },
               },
-              { section: 'progress', expectedEntries: { Phase: '8/8' } },
-            ],
-          } as AnalyticsTableRowDetails,
-        },
-      },
-      {
-        suiteTitle: 'with kuery query',
-        jobType: 'classification',
-        jobId: `fq_saved_search_3_${dateNow}`,
-        jobDescription: 'Classification job based on a saved search with kuery query',
-        source: 'ft_farequote_kuery',
-        get destinationIndex(): string {
-          return `user-${this.jobId}`;
-        },
-        runtimeFields: {
-          uppercase_airline: {
-            type: 'keyword',
-            script: 'emit(params._source.airline.toUpperCase())',
-          },
-        },
-        dependentVariable: 'airline',
-        trainingPercent: 20,
-        modelMemory: '20mb',
-        createIndexPattern: true,
-        expected: {
-          source: 'ft_farequote_small',
-          rocCurveColorState: [
-            // tick/grid/axis
-            { color: '#DDDDDD', percentage: 38 },
-            // line
-            { color: '#98A2B3', percentage: 7 },
-          ],
-          runtimeFieldsEditorContent: ['{', '  "uppercase_airline": {', '    "type": "keyword",'],
-          row: {
-            memoryStatus: 'ok',
-            type: 'classification',
-            status: 'stopped',
-            progress: '100',
-          },
-          rowDetails: {
-            jobDetails: [
-              {
-                section: 'state',
-                expectedEntries: {
-                  id: `fq_saved_search_3_${dateNow}`,
-                  state: 'stopped',
-                  data_counts:
-                    '{"training_docs_count":320,"test_docs_count":1283,"skipped_docs_count":0}',
-                  description: 'Classification job based on a saved search with kuery query',
-                },
-              },
-              { section: 'progress', expectedEntries: { Phase: '8/8' } },
+              { section: 'progress', expectedEntries: { Phase: '4/4' } },
             ],
           } as AnalyticsTableRowDetails,
         },
@@ -147,8 +118,6 @@ export default function ({ getService }: FtrProviderContext) {
           await ml.testExecution.logTestStep('loads the data frame analytics page');
           await ml.navigation.navigateToMl();
           await ml.navigation.navigateToDataFrameAnalytics();
-
-          await ml.testExecution.logTestStep('loads the source selection modal');
 
           // Disable anti-aliasing to stabilize canvas image rendering assertions
           await ml.commonUI.disableAntiAliasing();
@@ -183,19 +152,38 @@ export default function ({ getService }: FtrProviderContext) {
             testData.expected.runtimeFieldsEditorContent
           );
 
-          await ml.testExecution.logTestStep('inputs the dependent variable');
-          await ml.dataFrameAnalyticsCreation.assertDependentVariableInputExists();
-          await ml.dataFrameAnalyticsCreation.selectDependentVariable(testData.dependentVariable);
+          await ml.testExecution.logTestStep('does not display the dependent variable input');
+          await ml.dataFrameAnalyticsCreation.assertDependentVariableInputMissing();
 
-          await ml.testExecution.logTestStep('inputs the training percent');
-          await ml.dataFrameAnalyticsCreation.assertTrainingPercentInputExists();
-          await ml.dataFrameAnalyticsCreation.setTrainingPercent(testData.trainingPercent);
+          await ml.testExecution.logTestStep('does not display the training percent input');
+          await ml.dataFrameAnalyticsCreation.assertTrainingPercentInputMissing();
 
           await ml.testExecution.logTestStep('displays the source data preview');
           await ml.dataFrameAnalyticsCreation.assertSourceDataPreviewExists();
+          await ml.dataFrameAnalyticsCreation.assertSourceDataPreviewHistogramChartEnabled(true);
+
+          await ml.testExecution.logTestStep('displays the source data preview histogram charts');
+          await ml.dataFrameAnalyticsCreation.enableAndAssertSourceDataPreviewHistogramCharts(
+            testData.expected.histogramCharts
+          );
 
           await ml.testExecution.logTestStep('displays the include fields selection');
           await ml.dataFrameAnalyticsCreation.assertIncludeFieldsSelectionExists();
+
+          await ml.testExecution.logTestStep(
+            'sets the sample size to 10000 for the scatterplot matrix'
+          );
+          await ml.dataFrameAnalyticsCreation.setScatterplotMatrixSampleSizeSelectValue('10000');
+
+          await ml.testExecution.logTestStep(
+            'sets the randomize query switch to true for the scatterplot matrix'
+          );
+          await ml.dataFrameAnalyticsCreation.setScatterplotMatrixRandomizeQueryCheckState(true);
+
+          await ml.testExecution.logTestStep('displays the scatterplot matrix');
+          await ml.dataFrameAnalyticsCreation.assertScatterplotMatrix(
+            testData.expected.scatterplotMatrixColorsWizard
+          );
 
           await ml.testExecution.logTestStep('continues to the additional options step');
           await ml.dataFrameAnalyticsCreation.continueToAdditionalOptionsStep();
@@ -231,12 +219,7 @@ export default function ({ getService }: FtrProviderContext) {
 
           await ml.testExecution.logTestStep('checks validation callouts exist');
           await ml.dataFrameAnalyticsCreation.assertValidationCalloutsExists();
-          // Expect the follow callouts:
-          // - ✓ Dependent variable
-          // - ✓ Training percent
-          // - ✓ Top classes
-          // - ⚠ Analysis fields
-          await ml.dataFrameAnalyticsCreation.assertAllValidationCalloutsPresent(4);
+          await ml.dataFrameAnalyticsCreation.assertAllValidationCalloutsPresent(1);
 
           await ml.testExecution.logTestStep('continues to the create step');
           await ml.dataFrameAnalyticsCreation.continueToCreateStep();
@@ -275,7 +258,7 @@ export default function ({ getService }: FtrProviderContext) {
             id: testData.jobId,
             description: testData.jobDescription,
             memoryStatus: testData.expected.row.memoryStatus,
-            sourceIndex: testData.expected.source,
+            sourceIndex: testData.source,
             destinationIndex: testData.destinationIndex,
             type: testData.expected.row.type,
             status: testData.expected.row.status,
@@ -314,7 +297,7 @@ export default function ({ getService }: FtrProviderContext) {
             id: testData.jobId,
             description: editedDescription,
             memoryStatus: testData.expected.row.memoryStatus,
-            sourceIndex: testData.expected.source,
+            sourceIndex: testData.source,
             destinationIndex: testData.destinationIndex,
             type: testData.expected.row.type,
             status: testData.expected.row.status,
@@ -329,22 +312,24 @@ export default function ({ getService }: FtrProviderContext) {
 
           await ml.testExecution.logTestStep('displays the results view for created job');
           await ml.dataFrameAnalyticsTable.openResultsView(testData.jobId);
-          await ml.dataFrameAnalyticsResults.assertClassificationEvaluatePanelElementsExists();
-          await ml.dataFrameAnalyticsResults.assertClassificationTablePanelExists();
+          await ml.dataFrameAnalyticsResults.assertOutlierTablePanelExists();
           await ml.dataFrameAnalyticsResults.assertResultsTableExists();
-          await ml.dataFrameAnalyticsResults.assertResultsTableTrainingFiltersExist();
           await ml.dataFrameAnalyticsResults.assertResultsTableNotEmpty();
+          await ml.dataFrameAnalyticsResults.assertFeatureInfluenceCellNotEmpty();
 
-          await ml.testExecution.logTestStep('displays the ROC curve chart');
-          await ml.commonUI.assertColorsInCanvasElement(
-            'mlDFAnalyticsClassificationExplorationRocCurveChart',
-            testData.expected.rocCurveColorState,
-            ['#000000'],
-            undefined,
-            undefined,
-            // increased tolerance for ROC curve chart up from 10 to 20
-            // since the returned colors vary quite a bit on each run.
-            20
+          await ml.testExecution.logTestStep(
+            'sets the sample size to 10000 for the scatterplot matrix'
+          );
+          await ml.dataFrameAnalyticsResults.setScatterplotMatrixSampleSizeSelectValue('10000');
+
+          await ml.testExecution.logTestStep(
+            'sets the randomize query switch to true for the scatterplot matrix'
+          );
+          await ml.dataFrameAnalyticsResults.setScatterplotMatrixRandomizeQueryCheckState(true);
+
+          await ml.testExecution.logTestStep('displays the scatterplot matrix');
+          await ml.dataFrameAnalyticsResults.assertScatterplotMatrix(
+            testData.expected.scatterplotMatrixColorStatsResults
           );
 
           await ml.commonUI.resetAntiAliasing();
