+++ x-pack/test/functional/apps/ml/data_frame_analytics/results_view_content.ts
@@ -8,7 +8,7 @@
 import { DeepPartial } from '@kbn/ml-plugin/common/types/common';
 import { DataFrameAnalyticsConfig } from '@kbn/ml-plugin/public/application/data_frame_analytics/common';
 
-import { FtrProviderContext } from '../../../../ftr_provider_context';
+import { FtrProviderContext } from '../../../ftr_provider_context';
 
 export default function ({ getService }: FtrProviderContext) {
   const esArchiver = getService('esArchiver');
